# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: OpsPilot API
  version: 0.1.0
servers:
  - url: https://api.opspilot.com
  - url: http://localhost:4000
paths:
...
  /v1/actions/{id}/approve:
    post:
      summary: Approve action
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Action approved
  /v1/actions/{id}/execute:
    post:
      summary: Execute action
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Action queued
...